+++ activities/Memorize.activity/js/memorize-app.js
@@ -491,32 +491,18 @@ define(["activity/sample-ressources", "activity/palettes/template-palette", "act
                     MemorizeApp.context.decodeAudioData(b64, function (buffer) {
                         var source = MemorizeApp.context.createBufferSource(); // creates a sound source
                         if (MemorizeApp.source) {
-                            if (/iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream) {
-                                try {
-                                    MemorizeApp.source.noteOff(0);
-                                } catch (e) {
-                                }
-                            } else {
-                                try {
-                                    MemorizeApp.source.stop(0);
-                                } catch (e) {
-                                }
+                            try {
+                                MemorizeApp.source.stop(0);
+                            } catch (e) {
                             }
                         }
                         MemorizeApp.source = source;
                         source.buffer = buffer;
                         source.connect(MemorizeApp.context.destination);
-                        if (/iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream) {
-                            try {
-                                source.noteOn(0);
-                            } catch (e) {
-                            }
-                        } else {
-                            try {
-                                source.start(0);
-                            } catch (e) {
+                        try {
+                            source.start(0);
+                        } catch (e) {
 
-                            }
                         }
                     }, function (err) {
                         console.log("err(decodeAudioData): " + err);
@@ -658,7 +644,7 @@ define(["activity/sample-ressources", "activity/palettes/template-palette", "act
 
                 var clickEvent = "click";
                 if (/iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream) {
-                    clickEvent = "touchstart";
+                    clickEvent = "touchend";
                 }
                 fullCardDiv.addEventListener(clickEvent, onCardClick, false);
                 fullCardDiv.appendChild(div);